+++ dev-src/dev.clj
@@ -51,13 +51,11 @@
   (swap! b update :new-field inc)
   (swap! b dissoc :foo-field)
   
-  
 
+  ;; #trace (let [a {:a 10 :b [1 2 3]}
+  ;;              b 100]
+  ;;          (+ (:a a) b))
 
-  #trace (let [a {:a 10 :b [1 2 3]}
-               b 100]
-           (+ (:a a) b))
-
-  #trace (+ 1 (:a {:a 1}))
+  ;; #trace (+ 1 (:a {:a 1}))
   )